% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2instancespkg.R
\docType{package}
\name{ec2instances.info-package}
\alias{ec2instances.info-package}
\title{ec2 instances package}
\description{
This package contains a script to download data on Amazon EC2 pricing
from \link{ec2instances.info} and helper functions to format it nicely.
}
\examples{

#data(ec2instances)
plot(memory~vCPU, ec2instances)

}
\seealso{
\code{\link{ec2instances}}
}
