% Generated by roxygen2: do not edit by hand
% Please edit documentation in data/ec2instances.R
\docType{data}
\name{ec2instances}
\alias{ec2instances}
\title{ec2 Instances}
\format{An object of class \code{data.frame} with 76 rows and 23 columns.}
\source{
This data can be viewed on the web at \link{https://aws.amazon.com/ec2/pricing/on-demand/} and
 \link{https://aws.amazon.com/ec2/instance-types/}
}
\usage{
ec2instances
}
\description{
This data frame lists the different types of instances currently availabe on Amazon EC2.
}
\details{
\describe{
\item{ECU}{}
\item{arch}{i386 and/or x86_64}
\item{base_performance}{}
\item{burst_minutes}{}
\item{ebs_iops}{}
\item{ebs_max_bandwidth}{}
\item{ebs_optimized}{}
\item{ebs_throughput}{}
\item{enhanced_networking}{}
\item{family}{}
\item{generation}{previous or current}
\item{instance_type}{The short name, eg "m1.small"}
\item{ipv6_support}{}
\item{linux_virtualization_types}{PV and/or HVM}
\item{memory}{Gigabytes of memory}
\item{network_performance}{}
\item{placement_group_support}{}
\item{pretty_name}{eg  "M1 General Purpose Small"}
\item{pricing}{A nested data.frame containg prices for each AWS,
             region, OS, and on-demand or reservation contract}
\item{storage}{A nested data.frame containing storage size and type.}
\item{vCPU}{Number of virtual CPUs.}
\item{vpc}{A nested data.frame containg ips_per_eni and max_enis}
\item{vpc_only}{}
}
}
\keyword{datasets}
